requirements.txt
alembic==1.4.3
autopep8==1.5.4
Babel==2.9.0
blinker==1.4
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Email==1.4.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyMySQL==1.0.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
validate-docbr==1.8.2
Werkzeug==1.0.1
WTForms==2.3.3
flask-user==1.0.2.2
Flask-Toastr==0.5.4

